pyproject.toml
[project]
name = "simplemediasocial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.13",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
